+++ test/unit/mock/services/translate.js
@@ -1,0 +1,8 @@
+module.exports = [function () {
+    return {
+        $get: function () {},
+        then: function (successCallback) {
+            successCallback();
+        }
+    };
+}];
